Reorganize `internal/app/forklift` files/code
