+++ internal/app/forklift/caching-models.go
@@ -1,0 +1,21 @@
+package forklift
+
+import (
+	"github.com/PlanktoScope/forklift/pkg/pallets"
+)
+
+type CachedRepo struct {
+	VCSRepoPath string
+	Version     string
+	RepoSubdir  string
+	ConfigPath  string
+	Config      pallets.RepoConfig
+}
+
+type CachedPkg struct {
+	Repo       CachedRepo
+	Path       string
+	PkgSubdir  string
+	ConfigPath string
+	Config     pallets.PkgConfig
+}
