+++ internal/app/forklift/env-models.go
@@ -1,0 +1,10 @@
+// Package forklift provides the core functionality of the forklift tool
+package forklift
+
+type EnvSpec struct {
+	Description string `yaml:"description,omitempty"`
+}
+
+type EnvConfig struct {
+	Environment EnvSpec `yaml:"environment,omitempty"`
+}
