+++ internal/app/forklift/external-models.go
@@ -1,0 +1,10 @@
+package forklift
+
+import (
+	"io/fs"
+)
+
+type ExternalRepo struct {
+	Repo CachedRepo
+	FS   fs.FS
+}
