+++ internal/app/forklift/external-pkgs.go
@@ -11,77 +11,7 @@ import (
 	"github.com/PlanktoScope/forklift/pkg/pallets"
 )
 
-func ListExternalPkgs(repo ExternalRepo, cachedPrefix string) ([]CachedPkg, error) {
-	searchPattern := fmt.Sprintf("**/%s", pallets.PkgSpecFile)
-	if cachedPrefix != "" {
-		searchPattern = filepath.Join(cachedPrefix, searchPattern)
-	}
-	pkgConfigFiles, err := doublestar.Glob(repo.FS, searchPattern)
-	if err != nil {
-		return nil, errors.Wrap(err, "couldn't search for external package configs")
-	}
-
-	repoPkgPaths := make([]string, 0, len(pkgConfigFiles))
-	pkgMap := make(map[string]CachedPkg)
-	for _, pkgConfigFilePath := range pkgConfigFiles {
-		filename := filepath.Base(pkgConfigFilePath)
-		if filename != pallets.PkgSpecFile {
-			continue
-		}
-		pkg, err := loadExternalPkg(repo, pkgConfigFilePath)
-		if err != nil {
-			return nil, errors.Wrapf(err, "couldn't load cached package from %s", pkgConfigFilePath)
-		}
-
-		if prevPkg, ok := pkgMap[pkg.Path]; ok {
-			if prevPkg.Repo.FromSameVCSRepo(pkg.Repo) && prevPkg.ConfigPath != pkg.ConfigPath {
-				return nil, errors.Errorf(
-					"package repeatedly defined in the same version of the same Github repo: %s, %s",
-					prevPkg.ConfigPath, pkg.ConfigPath,
-				)
-			}
-		}
-		repoPkgPaths = append(repoPkgPaths, pkg.Path)
-		pkgMap[pkg.Path] = pkg
-	}
-
-	orderedPkgs := make([]CachedPkg, 0, len(repoPkgPaths))
-	for _, path := range repoPkgPaths {
-		orderedPkgs = append(orderedPkgs, pkgMap[path])
-	}
-	return orderedPkgs, nil
-}
-
-func loadExternalPkg(repo ExternalRepo, pkgConfigFilePath string) (CachedPkg, error) {
-	config, err := loadPkgConfig(repo.FS, pkgConfigFilePath)
-	if err != nil {
-		return CachedPkg{}, errors.Wrapf(
-			err, "couldn't load external package config from %s", pkgConfigFilePath,
-		)
-	}
-
-	pkg := CachedPkg{
-		Repo:       repo.Repo,
-		ConfigPath: filepath.Join(repo.Repo.ConfigPath, filepath.Dir(pkgConfigFilePath)),
-		Config:     config,
-	}
-	pkg.PkgSubdir = strings.TrimPrefix(pkg.ConfigPath, fmt.Sprintf("%s/", pkg.Repo.ConfigPath))
-	pkg.Path = fmt.Sprintf("%s/%s", pkg.Repo.Config.Repository.Path, pkg.PkgSubdir)
-	return pkg, nil
-}
-
-func FindExternalRepoOfPkg(
-	repos map[string]ExternalRepo, pkgPath string,
-) (repo ExternalRepo, ok bool) {
-	repoCandidatePath := filepath.Dir(pkgPath)
-	for repoCandidatePath != "." {
-		if repo, ok = repos[repoCandidatePath]; ok {
-			return repo, true
-		}
-		repoCandidatePath = filepath.Dir(repoCandidatePath)
-	}
-	return ExternalRepo{}, false
-}
+// Loading
 
 func FindExternalPkg(repo ExternalRepo, pkgPath string) (CachedPkg, error) {
 	pkgInnermostDir := filepath.Base(pkgPath)
@@ -126,6 +56,24 @@ func FindExternalPkg(repo ExternalRepo, pkgPath string) (CachedPkg, error) {
 	return candidatePkgs[0], nil
 }
 
+func loadExternalPkg(repo ExternalRepo, pkgConfigFilePath string) (CachedPkg, error) {
+	config, err := loadPkgConfig(repo.FS, pkgConfigFilePath)
+	if err != nil {
+		return CachedPkg{}, errors.Wrapf(
+			err, "couldn't load external package config from %s", pkgConfigFilePath,
+		)
+	}
+
+	pkg := CachedPkg{
+		Repo:       repo.Repo,
+		ConfigPath: filepath.Join(repo.Repo.ConfigPath, filepath.Dir(pkgConfigFilePath)),
+		Config:     config,
+	}
+	pkg.PkgSubdir = strings.TrimPrefix(pkg.ConfigPath, fmt.Sprintf("%s/", pkg.Repo.ConfigPath))
+	pkg.Path = fmt.Sprintf("%s/%s", pkg.Repo.Config.Repository.Path, pkg.PkgSubdir)
+	return pkg, nil
+}
+
 func AsVersionedPkg(pkg CachedPkg) VersionedPkg {
 	return VersionedPkg{
 		Path: pkg.Path,
@@ -135,4 +83,47 @@ func AsVersionedPkg(pkg CachedPkg) VersionedPkg {
 		},
 		Cached: pkg,
 	}
+}
+
+// Listing
+
+func ListExternalPkgs(repo ExternalRepo, cachedPrefix string) ([]CachedPkg, error) {
+	searchPattern := fmt.Sprintf("**/%s", pallets.PkgSpecFile)
+	if cachedPrefix != "" {
+		searchPattern = filepath.Join(cachedPrefix, searchPattern)
+	}
+	pkgConfigFiles, err := doublestar.Glob(repo.FS, searchPattern)
+	if err != nil {
+		return nil, errors.Wrap(err, "couldn't search for external package configs")
+	}
+
+	repoPkgPaths := make([]string, 0, len(pkgConfigFiles))
+	pkgMap := make(map[string]CachedPkg)
+	for _, pkgConfigFilePath := range pkgConfigFiles {
+		filename := filepath.Base(pkgConfigFilePath)
+		if filename != pallets.PkgSpecFile {
+			continue
+		}
+		pkg, err := loadExternalPkg(repo, pkgConfigFilePath)
+		if err != nil {
+			return nil, errors.Wrapf(err, "couldn't load cached package from %s", pkgConfigFilePath)
+		}
+
+		if prevPkg, ok := pkgMap[pkg.Path]; ok {
+			if prevPkg.Repo.FromSameVCSRepo(pkg.Repo) && prevPkg.ConfigPath != pkg.ConfigPath {
+				return nil, errors.Errorf(
+					"package repeatedly defined in the same version of the same Github repo: %s, %s",
+					prevPkg.ConfigPath, pkg.ConfigPath,
+				)
+			}
+		}
+		repoPkgPaths = append(repoPkgPaths, pkg.Path)
+		pkgMap[pkg.Path] = pkg
+	}
+
+	orderedPkgs := make([]CachedPkg, 0, len(repoPkgPaths))
+	for _, path := range repoPkgPaths {
+		orderedPkgs = append(orderedPkgs, pkgMap[path])
+	}
+	return orderedPkgs, nil
 }
