+++ internal/app/forklift/versioning-pkgs.go
@@ -8,6 +8,52 @@ import (
 	"github.com/pkg/errors"
 )
 
+// Loading
+
+func LoadVersionedPkg(reposFS, cacheFS fs.FS, pkgPath string) (VersionedPkg, error) {
+	repo, err := findVersionedRepoOfPkg(reposFS, pkgPath)
+	if err != nil {
+		return VersionedPkg{}, errors.Wrapf(
+			err, "couldn't find repo providing package %s in local environment", pkgPath,
+		)
+	}
+	version, err := repo.Config.Version()
+	if err != nil {
+		return VersionedPkg{}, errors.Wrapf(
+			err, "couldn't determine version of repo %s in local environment", repo.Path(),
+		)
+	}
+	pkg, err := FindCachedPkg(cacheFS, pkgPath, version)
+	if err != nil {
+		return VersionedPkg{}, errors.Wrapf(
+			err, "couldn't find package %s@%s in cache", pkgPath, version,
+		)
+	}
+
+	return VersionedPkg{
+		Path:   pkgPath,
+		Repo:   repo,
+		Cached: pkg,
+	}, nil
+}
+
+func findVersionedRepoOfPkg(reposFS fs.FS, pkgPath string) (VersionedRepo, error) {
+	repoCandidatePath := filepath.Dir(pkgPath)
+	for repoCandidatePath != "." {
+		repo, err := LoadVersionedRepo(reposFS, repoCandidatePath)
+		if err == nil {
+			return repo, nil
+		}
+		repoCandidatePath = filepath.Dir(repoCandidatePath)
+	}
+	return VersionedRepo{}, errors.Errorf(
+		"no repository config file found in %s or any parent directory in local environment",
+		filepath.Dir(pkgPath),
+	)
+}
+
+// Listing
+
 func ListVersionedPkgs(
 	cacheFS fs.FS, replacementRepos map[string]ExternalRepo, repos []VersionedRepo,
 ) (orderedPkgs []CachedPkg, err error) {
@@ -101,46 +147,4 @@ func listVersionedPkgsOfCachedRepo(
 	}
 
 	return pkgMap, versionedPkgPaths, nil
-}
-
-func LoadVersionedPkg(reposFS, cacheFS fs.FS, pkgPath string) (VersionedPkg, error) {
-	repo, err := findVersionedRepoOfPkg(reposFS, pkgPath)
-	if err != nil {
-		return VersionedPkg{}, errors.Wrapf(
-			err, "couldn't find repo providing package %s in local environment", pkgPath,
-		)
-	}
-	version, err := repo.Config.Version()
-	if err != nil {
-		return VersionedPkg{}, errors.Wrapf(
-			err, "couldn't determine version of repo %s in local environment", repo.Path(),
-		)
-	}
-	pkg, err := FindCachedPkg(cacheFS, pkgPath, version)
-	if err != nil {
-		return VersionedPkg{}, errors.Wrapf(
-			err, "couldn't find package %s@%s in cache", pkgPath, version,
-		)
-	}
-
-	return VersionedPkg{
-		Path:   pkgPath,
-		Repo:   repo,
-		Cached: pkg,
-	}, nil
-}
-
-func findVersionedRepoOfPkg(reposFS fs.FS, pkgPath string) (VersionedRepo, error) {
-	repoCandidatePath := filepath.Dir(pkgPath)
-	for repoCandidatePath != "." {
-		repo, err := LoadVersionedRepo(reposFS, repoCandidatePath)
-		if err == nil {
-			return repo, nil
-		}
-		repoCandidatePath = filepath.Dir(repoCandidatePath)
-	}
-	return VersionedRepo{}, errors.Errorf(
-		"no repository config file found in %s or any parent directory in local environment",
-		filepath.Dir(pkgPath),
-	)
 }
